cleanup

+++ src/gvar/_utilities.pyx
@@ -554,7 +554,7 @@ def evalcov_blocks(g, compress=False):
         if len(bl) == 1:
             b = bl.pop()
             gb = gf[b]
-            ans.append((numpy.array([b], dtype=numpy.intp), numpy.array([[gb.var]])))
+            ans.append((numpy.array([b]), numpy.array([[gb.var]])))
         else:
             idx = numpy.array([b for b in bl], dtype=numpy.intp)
             ans.append((idx, evalcov(gf[idx])))
@@ -566,7 +566,7 @@ def evalcov_blocks(g, compress=False):
                 c_ans[0][1].append(bcov[0, 0] ** 0.5)
             else:
                 c_ans.append((idx, bcov))
-        c_ans[0] = (numpy.array(c_ans[0][0], dtype=numpy.intp), numpy.array(c_ans[0][1], dtype=float))            
+        c_ans[0] = (numpy.array(c_ans[0][0], dtype=int), numpy.array(c_ans[0][1], dtype=float))            
         return c_ans
     else:
         return ans
